+++ esp-hal/src/dma/mod.rs
@@ -205,7 +205,6 @@ where
 bitfield::bitfield! {
     #[doc(hidden)]
     #[derive(Clone, Copy)]
-    #[cfg_attr(feature = "defmt", derive(defmt::Format))]
     pub struct DmaDescriptorFlags(u32);
 
     u16;
@@ -223,6 +222,20 @@ impl Debug for DmaDescriptorFlags {
             .field("suc_eof", &self.suc_eof())
             .field("owner", &(if self.owner() { "DMA" } else { "CPU" }))
             .finish()
+    }
+}
+
+#[cfg(feature = "defmt")]
+impl defmt::Format for DmaDescriptorFlags {
+    fn format(&self, fmt: defmt::Formatter<'_>) {
+        defmt::write!(
+            fmt,
+            "DmaDescriptorFlags {{ size: {}, length: {}, suc_eof: {}, owner: {} }}",
+            self.size(),
+            self.length(),
+            self.suc_eof(),
+            if self.owner() { "DMA" } else { "CPU" }
+        );
     }
 }
 
@@ -286,6 +299,8 @@ use enumset::{EnumSet, EnumSetType};
 pub use self::gdma::*;
 #[cfg(pdma)]
 pub use self::pdma::*;
+#[cfg(esp32s3)]
+use crate::soc::is_slice_in_psram;
 use crate::{interrupt::InterruptHandler, soc::is_slice_in_dram, Mode};
 
 #[cfg(gdma)]
@@ -486,18 +501,7 @@ macro_rules! as_mut_byte_array {
 #[macro_export]
 macro_rules! dma_buffers_chunk_size {
     ($rx_size:expr, $tx_size:expr, $chunk_size:expr) => {{
-        $crate::declare_aligned_dma_buffer!(RX_BUFFER, $rx_size);
-        $crate::declare_aligned_dma_buffer!(TX_BUFFER, $tx_size);
-        let (mut rx_descriptors, mut tx_descriptors) =
-            $crate::dma_descriptors_chunk_size!($rx_size, $tx_size, $chunk_size);
-        unsafe {
-            (
-                $crate::as_mut_byte_array!(RX_BUFFER, $rx_size),
-                rx_descriptors,
-                $crate::as_mut_byte_array!(TX_BUFFER, $tx_size),
-                tx_descriptors,
-            )
-        }
+        $crate::dma_buffers_impl!($rx_size, $tx_size, $chunk_size, is_circular = false)
     }};
 
     ($size:expr, $chunk_size:expr) => {
@@ -522,18 +526,7 @@ macro_rules! dma_buffers_chunk_size {
 #[macro_export]
 macro_rules! dma_circular_buffers_chunk_size {
     ($rx_size:expr, $tx_size:expr, $chunk_size:expr) => {{
-        $crate::declare_aligned_dma_buffer!(RX_BUFFER, $rx_size);
-        $crate::declare_aligned_dma_buffer!(TX_BUFFER, $tx_size);
-        let (mut rx_descriptors, mut tx_descriptors) =
-            $crate::dma_circular_descriptors_chunk_size!($rx_size, $tx_size, $chunk_size);
-        unsafe {
-            (
-                $crate::as_mut_byte_array!(RX_BUFFER, $rx_size),
-                rx_descriptors,
-                $crate::as_mut_byte_array!(TX_BUFFER, $tx_size),
-                tx_descriptors,
-            )
-        }
+        $crate::dma_buffers_impl!($rx_size, $tx_size, $chunk_size, is_circular = true)
     }};
 
     ($size:expr, $chunk_size:expr) => {{
@@ -557,17 +550,7 @@ macro_rules! dma_circular_buffers_chunk_size {
 #[macro_export]
 macro_rules! dma_descriptors_chunk_size {
     ($rx_size:expr, $tx_size:expr, $chunk_size:expr) => {{
-        // these will check for size at compile time
-        const _: () = ::core::assert!($chunk_size <= 4092, "chunk size must be <= 4092");
-        const _: () = ::core::assert!($chunk_size > 0, "chunk size must be > 0");
-
-        static mut RX_DESCRIPTORS: [$crate::dma::DmaDescriptor;
-            ($rx_size + $chunk_size - 1) / $chunk_size] =
-            [$crate::dma::DmaDescriptor::EMPTY; ($rx_size + $chunk_size - 1) / $chunk_size];
-        static mut TX_DESCRIPTORS: [$crate::dma::DmaDescriptor;
-            ($tx_size + $chunk_size - 1) / $chunk_size] =
-            [$crate::dma::DmaDescriptor::EMPTY; ($tx_size + $chunk_size - 1) / $chunk_size];
-        unsafe { (&mut RX_DESCRIPTORS, &mut TX_DESCRIPTORS) }
+        $crate::dma_descriptors_impl!($rx_size, $tx_size, $chunk_size, is_circular = false)
     }};
 
     ($size:expr, $chunk_size:expr) => {
@@ -592,32 +575,103 @@ macro_rules! dma_descriptors_chunk_size {
 #[macro_export]
 macro_rules! dma_circular_descriptors_chunk_size {
     ($rx_size:expr, $tx_size:expr, $chunk_size:expr) => {{
-        // these will check for size at compile time
-        const _: () = ::core::assert!($chunk_size <= 4092, "chunk size must be <= 4092");
-        const _: () = ::core::assert!($chunk_size > 0, "chunk size must be > 0");
-
-        const rx_descriptor_len: usize = if $rx_size > $chunk_size * 2 {
-            ($rx_size + $chunk_size - 1) / $chunk_size
-        } else {
-            3
-        };
-
-        const tx_descriptor_len: usize = if $tx_size > $chunk_size * 2 {
-            ($tx_size + $chunk_size - 1) / $chunk_size
-        } else {
-            3
-        };
-
-        static mut RX_DESCRIPTORS: [$crate::dma::DmaDescriptor; rx_descriptor_len] =
-            [$crate::dma::DmaDescriptor::EMPTY; rx_descriptor_len];
-        static mut TX_DESCRIPTORS: [$crate::dma::DmaDescriptor; tx_descriptor_len] =
-            [$crate::dma::DmaDescriptor::EMPTY; tx_descriptor_len];
-        unsafe { (&mut RX_DESCRIPTORS, &mut TX_DESCRIPTORS) }
+        $crate::dma_descriptors_impl!($rx_size, $tx_size, $chunk_size, is_circular = true)
     }};
 
     ($size:expr, $chunk_size:expr) => {
         $crate::dma_circular_descriptors_chunk_size!($size, $size, $chunk_size)
     };
+}
+
+#[doc(hidden)]
+#[macro_export]
+macro_rules! dma_buffers_impl {
+    ($rx_size:expr, $tx_size:expr, $chunk_size:expr, is_circular = $circular:tt) => {{
+        let rx = $crate::dma_buffers_impl!($rx_size, $chunk_size, is_circular = $circular);
+        let tx = $crate::dma_buffers_impl!($tx_size, $chunk_size, is_circular = $circular);
+        (rx.0, rx.1, tx.0, tx.1)
+    }};
+
+    ($size:expr, $chunk_size:expr, is_circular = $circular:tt) => {{
+        $crate::declare_aligned_dma_buffer!(BUFFER, $size);
+
+        unsafe {
+            (
+                $crate::as_mut_byte_array!(BUFFER, $size),
+                $crate::dma_descriptors_impl!($size, $chunk_size, is_circular = $circular),
+            )
+        }
+    }};
+}
+
+#[doc(hidden)]
+#[macro_export]
+macro_rules! dma_descriptors_impl {
+    ($rx_size:expr, $tx_size:expr, $chunk_size:expr, is_circular = $circular:tt) => {{
+        let rx = $crate::dma_descriptors_impl!($rx_size, $chunk_size, is_circular = $circular);
+        let tx = $crate::dma_descriptors_impl!($tx_size, $chunk_size, is_circular = $circular);
+        (rx, tx)
+    }};
+
+    ($size:expr, $chunk_size:expr, is_circular = $circular:tt) => {{
+        const COUNT: usize =
+            $crate::dma_descriptor_count!($size, $chunk_size, is_circular = $circular);
+
+        static mut DESCRIPTORS: [$crate::dma::DmaDescriptor; COUNT] =
+            [$crate::dma::DmaDescriptor::EMPTY; COUNT];
+
+        unsafe { &mut DESCRIPTORS }
+    }};
+}
+
+#[doc(hidden)]
+#[macro_export]
+macro_rules! dma_descriptor_count {
+    (@validate_chunk_size $chunk_size:expr) => {
+        const {
+            ::core::assert!($chunk_size <= 4095, "chunk size must be <= 4095");
+            ::core::assert!($chunk_size > 0, "chunk size must be > 0");
+        }
+    };
+
+    ($size:expr, $chunk_size:expr, is_circular = true) => {{
+        $crate::dma_descriptor_count!(@validate_chunk_size $chunk_size);
+        if $size > $chunk_size * 2 {
+            ($size as usize).div_ceil($chunk_size)
+        } else {
+            3
+        }
+    }};
+
+    ($size:expr, $chunk_size:expr, is_circular = false) => {{
+        $crate::dma_descriptor_count!(@validate_chunk_size $chunk_size);
+        ($size as usize).div_ceil($chunk_size)
+    }};
+}
+
+/// Convenience macro to create a DmaTxBuf from buffer size. The buffer and
+/// descriptors are statically allocated and used to create the `DmaTxBuf`.
+///
+/// ## Usage
+/// ```rust,no_run
+#[doc = crate::before_snippet!()]
+/// use esp_hal::dma_tx_buffer;
+/// use esp_hal::dma::DmaBufBlkSize;
+///
+/// let tx_buf = dma_tx_buffer!(32000);
+/// # }
+/// ```
+#[macro_export]
+macro_rules! dma_tx_buffer {
+    ($tx_size:expr) => {{
+        let (tx_buffer, tx_descriptors) = $crate::dma_buffers_impl!(
+            $tx_size,
+            $crate::dma::DmaTxBuf::compute_chunk_size(None),
+            is_circular = false
+        );
+
+        $crate::dma::DmaTxBuf::new(tx_descriptors, tx_buffer)
+    }};
 }
 
 /// DMA Errors
@@ -787,14 +841,6 @@ pub trait AesPeripheral: PeripheralMarker {}
 /// Marks channels as usable for LCD_CAM
 #[doc(hidden)]
 pub trait LcdCamPeripheral: PeripheralMarker {}
-
-/// DMA Rx
-#[doc(hidden)]
-pub trait Rx: RxPrivate {}
-
-/// DMA Tx
-#[doc(hidden)]
-pub trait Tx: TxPrivate {}
 
 /// Marker trait
 #[doc(hidden)]
@@ -1001,6 +1047,16 @@ pub enum DmaExtMemBKSize {
     Size64 = 2,
 }
 
+impl From<DmaBufBlkSize> for DmaExtMemBKSize {
+    fn from(size: DmaBufBlkSize) -> Self {
+        match size {
+            DmaBufBlkSize::Size16 => DmaExtMemBKSize::Size16,
+            DmaBufBlkSize::Size32 => DmaExtMemBKSize::Size32,
+            DmaBufBlkSize::Size64 => DmaExtMemBKSize::Size64,
+        }
+    }
+}
+
 pub(crate) struct TxCircularState {
     write_offset: usize,
     write_descr_ptr: *mut DmaDescriptor,
@@ -1028,7 +1084,7 @@ impl TxCircularState {
 
     pub(crate) fn update<T>(&mut self, channel: &T)
     where
-        T: TxPrivate,
+        T: Tx,
     {
         if channel
             .pending_out_interrupts()
@@ -1238,7 +1294,7 @@ impl RxCircularState {
 /// A description of a DMA Channel.
 pub trait DmaChannel: crate::private::Sealed {
     #[doc(hidden)]
-    type Channel: ChannelTypes + RegisterAccess;
+    type Channel: RegisterAccess;
 
     /// A description of the RX half of a DMA Channel.
     type Rx: RxChannel<Self::Channel>;
@@ -1248,11 +1304,14 @@ pub trait DmaChannel: crate::private::Sealed {
 
     /// A suitable peripheral for this DMA channel.
     type P: PeripheralMarker;
+
+    #[doc(hidden)]
+    fn set_isr(handler: InterruptHandler);
 }
 
 /// The functions here are not meant to be used outside the HAL
 #[doc(hidden)]
-pub trait RxPrivate: crate::private::Sealed {
+pub trait Rx: crate::private::Sealed {
     fn init(&mut self, burst_mode: bool, priority: DmaPriority);
 
     unsafe fn prepare_transfer_without_start(
@@ -1374,11 +1433,9 @@ where
     }
 }
 
-impl<'a, CH> Rx for ChannelRx<'a, CH> where CH: DmaChannel {}
-
 impl<'a, CH> crate::private::Sealed for ChannelRx<'a, CH> where CH: DmaChannel {}
 
-impl<'a, CH> RxPrivate for ChannelRx<'a, CH>
+impl<'a, CH> Rx for ChannelRx<'a, CH>
 where
     CH: DmaChannel,
 {
@@ -1414,7 +1471,6 @@ where
                 if des.buffer as usize % alignment != 0 && des.size() % alignment != 0 {
                     return Err(DmaError::InvalidAlignment);
                 }
-                // TODO: make this optional?
                 crate::soc::cache_invalidate_addr(des.buffer as u32, des.size() as u32);
             }
         }
@@ -1489,7 +1545,7 @@ where
 
 /// The functions here are not meant to be used outside the HAL
 #[doc(hidden)]
-pub trait TxPrivate: crate::private::Sealed {
+pub trait Tx: crate::private::Sealed {
     fn init(&mut self, burst_mode: bool, priority: DmaPriority);
 
     unsafe fn prepare_transfer_without_start(
@@ -1621,11 +1677,9 @@ where
     }
 }
 
-impl<'a, CH> Tx for ChannelTx<'a, CH> where CH: DmaChannel {}
-
 impl<'a, CH> crate::private::Sealed for ChannelTx<'a, CH> where CH: DmaChannel {}
 
-impl<'a, CH> TxPrivate for ChannelTx<'a, CH>
+impl<'a, CH> Tx for ChannelTx<'a, CH>
 where
     CH: DmaChannel,
 {
@@ -1638,6 +1692,7 @@ where
         peri: DmaPeripheral,
         chain: &DescriptorChain,
     ) -> Result<(), DmaError> {
+        // TODO: based on the ESP32-S3 TRM the alignment check is not needed for TX!
         // for esp32s3 we check each descriptor buffer that points to psram for
         // alignment and writeback the cache for that buffer
         #[cfg(esp32s3)]
@@ -1663,7 +1718,19 @@ where
         buffer: &mut BUF,
     ) -> Result<(), DmaError> {
         let preparation = buffer.prepare();
-
+        cfg_if::cfg_if!(
+            if #[cfg(esp32s3)] {
+                if let Some(block_size) = preparation.block_size {
+                    self.set_ext_mem_block_size(block_size.into());
+                }
+            } else {
+                // we insure that block_size is some only for PSRAM addresses
+                if preparation.block_size.is_some() {
+                    return Err(DmaError::UnsupportedMemoryRegion);
+                }
+            }
+        );
+        // TODO: Get burst mode from DmaBuf.
         self.tx_impl
             .prepare_transfer_without_start(preparation.start, peri)
     }
@@ -1754,11 +1821,6 @@ pub trait RegisterAccess: crate::private::Sealed {
     fn start_in();
 }
 
-#[doc(hidden)]
-pub trait ChannelTypes: crate::private::Sealed {
-    fn set_isr(handler: InterruptHandler);
-}
-
 /// DMA Channel
 pub struct Channel<'d, CH, MODE>
 where
@@ -1781,7 +1843,7 @@ where
     ///
     /// Interrupts are not enabled at the peripheral level here.
     pub fn set_interrupt_handler(&mut self, handler: InterruptHandler) {
-        <C::Channel as ChannelTypes>::set_isr(handler);
+        C::set_isr(handler);
     }
 
     /// Listen for the given interrupts
@@ -1830,6 +1892,10 @@ where
 /// Holds all the information needed to configure a DMA channel for a transfer.
 pub struct Preparation {
     start: *mut DmaDescriptor,
+    /// block size for PSRAM transfers (TODO: enable burst mode for non external
+    /// memory?)
+    #[cfg_attr(not(esp32s3), allow(dead_code))]
+    block_size: Option<DmaBufBlkSize>,
     // burst_mode, alignment, check_owner, etc.
 }
 
@@ -1872,22 +1938,41 @@ pub trait DmaRxBuffer {
 
 /// Error returned from Dma[Rx|Tx|RxTx]Buf operations.
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum DmaBufError {
     /// More descriptors are needed for the buffer size
     InsufficientDescriptors,
     /// Descriptors or buffers are not located in a supported memory region
     UnsupportedMemoryRegion,
+    /// Buffer is not aligned to the required size
+    InvalidAlignment,
+    /// Invalid chunk size: must be > 0 and <= 4095
+    InvalidChunkSize,
+}
+
+/// DMA buffer alignments
+#[derive(Debug, Clone, Copy, PartialEq)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum DmaBufBlkSize {
+    /// 16 bytes
+    Size16 = 16,
+    /// 32 bytes
+    Size32 = 32,
+    /// 64 bytes
+    Size64 = 64,
 }
 
 /// DMA transmit buffer
 ///
 /// This is a contiguous buffer linked together by DMA descriptors of length
-/// 4092. It can only be used for transmitting data to a peripheral's FIFO.
-/// See [DmaRxBuf] for receiving data.
+/// 4095 at most. It can only be used for transmitting data to a peripheral's
+/// FIFO. See [DmaRxBuf] for receiving data.
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct DmaTxBuf {
     descriptors: &'static mut [DmaDescriptor],
     buffer: &'static mut [u8],
+    block_size: Option<DmaBufBlkSize>,
 }
 
 impl DmaTxBuf {
@@ -1897,23 +1982,87 @@ impl DmaTxBuf {
     /// Each descriptor can handle 4092 bytes worth of buffer.
     ///
     /// Both the descriptors and buffer must be in DMA-capable memory.
-    /// Only DRAM is supported.
+    /// Only DRAM is supported for descriptors.
     pub fn new(
         descriptors: &'static mut [DmaDescriptor],
         buffer: &'static mut [u8],
     ) -> Result<Self, DmaBufError> {
-        let min_descriptors = buffer.len().div_ceil(CHUNK_SIZE);
+        Self::new_with_block_size(descriptors, buffer, None)
+    }
+
+    /// Compute max chunk size based on block size
+    pub const fn compute_chunk_size(block_size: Option<DmaBufBlkSize>) -> usize {
+        match block_size {
+            Some(size) => 4096 - size as usize,
+            #[cfg(esp32)]
+            None => 4092, // esp32 requires 4 byte alignment
+            #[cfg(not(esp32))]
+            None => 4095,
+        }
+    }
+
+    /// Compute the number of descriptors required for a given block size and
+    /// buffer size
+    pub const fn compute_descriptor_count(
+        buffer_size: usize,
+        block_size: Option<DmaBufBlkSize>,
+    ) -> usize {
+        buffer_size.div_ceil(Self::compute_chunk_size(block_size))
+    }
+
+    /// Creates a new [DmaTxBuf] from some descriptors and a buffer.
+    ///
+    /// There must be enough descriptors for the provided buffer.
+    /// Each descriptor can handle at most 4095 bytes worth of buffer.
+    /// Optionally, a block size can be provided for PSRAM & Burst transfers.
+    ///
+    /// Both the descriptors and buffer must be in DMA-capable memory.
+    /// Only DRAM is supported for descriptors.
+    pub fn new_with_block_size(
+        descriptors: &'static mut [DmaDescriptor],
+        buffer: &'static mut [u8],
+        block_size: Option<DmaBufBlkSize>,
+    ) -> Result<Self, DmaBufError> {
+        let chunk_size = Self::compute_chunk_size(block_size);
+        let min_descriptors = Self::compute_descriptor_count(buffer.len(), block_size);
         if descriptors.len() < min_descriptors {
             return Err(DmaBufError::InsufficientDescriptors);
         }
 
-        if !is_slice_in_dram(descriptors) || !is_slice_in_dram(buffer) {
+        // descriptors are required to be in DRAM
+        if !is_slice_in_dram(descriptors) {
             return Err(DmaBufError::UnsupportedMemoryRegion);
+        }
+
+        cfg_if::cfg_if! {
+            if #[cfg(esp32s3)] {
+                // buffer can be either DRAM or PSRAM (if supported)
+                if !is_slice_in_dram(buffer) && !is_slice_in_psram(buffer) {
+                    return Err(DmaBufError::UnsupportedMemoryRegion);
+                }
+                // if its PSRAM, the block_size/alignment must be specified
+                if is_slice_in_psram(buffer) && block_size.is_none() {
+                    return Err(DmaBufError::InvalidAlignment);
+                }
+            } else {
+                #[cfg(any(esp32,esp32s2))]
+                if buffer.len() % 4 != 0 && buffer.as_ptr() as usize % 4 != 0 {
+                    // ESP32 requires word alignment for DMA buffers.
+                    // ESP32-S2 technically supports byte-aligned DMA buffers, but the
+                    // transfer ends up writing out of bounds if the buffer's length
+                    // is 2 or 3 (mod 4).
+                    return Err(DmaBufError::InvalidAlignment);
+                }
+                // buffer can only be DRAM
+                if !is_slice_in_dram(buffer) {
+                    return Err(DmaBufError::UnsupportedMemoryRegion);
+                }
+            }
         }
 
         // Setup size and buffer pointer as these will not change for the remainder of
         // this object's lifetime
-        let chunk_iter = descriptors.iter_mut().zip(buffer.chunks_mut(CHUNK_SIZE));
+        let chunk_iter = descriptors.iter_mut().zip(buffer.chunks_mut(chunk_size));
         for (desc, chunk) in chunk_iter {
             desc.set_size(chunk.len());
             desc.buffer = chunk.as_mut_ptr();
@@ -1922,9 +2071,13 @@ impl DmaTxBuf {
         let mut buf = Self {
             descriptors,
             buffer,
+            block_size,
         };
         buf.set_length(buf.capacity());
-
+        // no need for block size if the buffer is in DRAM
+        if is_slice_in_dram(buf.buffer) {
+            buf.block_size = None;
+        }
         Ok(buf)
     }
 
@@ -1960,7 +2113,7 @@ impl DmaTxBuf {
         assert!(len <= self.buffer.len());
 
         // Get the minimum number of descriptors needed for this length of data.
-        let descriptor_count = len.div_ceil(CHUNK_SIZE).max(1);
+        let descriptor_count = len.div_ceil(self.descriptors[0].size()).max(1);
         let required_descriptors = &mut self.descriptors[0..descriptor_count];
 
         // Link up the relevant descriptors.
@@ -2021,8 +2174,19 @@ impl DmaTxBuffer for DmaTxBuf {
             }
         }
 
+        #[cfg(esp32s3)]
+        if crate::soc::is_valid_psram_address(self.buffer.as_ptr() as u32) {
+            unsafe {
+                crate::soc::cache_writeback_addr(
+                    self.buffer.as_ptr() as u32,
+                    self.buffer.len() as u32,
+                )
+            };
+        }
+
         Preparation {
             start: self.descriptors.as_mut_ptr(),
+            block_size: self.block_size,
         }
     }
 
@@ -2259,6 +2423,7 @@ impl DmaRxBuffer for DmaRxBuf {
 
         Preparation {
             start: self.descriptors.as_mut_ptr(),
+            block_size: None,
         }
     }
 
@@ -2451,6 +2616,7 @@ impl DmaTxBuffer for DmaRxTxBuf {
 
         Preparation {
             start: self.tx_descriptors.as_mut_ptr(),
+            block_size: None, // TODO: support block size!
         }
     }
 
@@ -2480,6 +2646,7 @@ impl DmaRxBuffer for DmaRxTxBuf {
 
         Preparation {
             start: self.rx_descriptors.as_mut_ptr(),
+            block_size: None, // TODO: support block size!
         }
     }