+++ hil-test/tests/qspi.rs
@@ -1,7 +1,6 @@
 //! QSPI Test Suite
 
 //% CHIPS: esp32 esp32c2 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-//% FEATURES: defmt
 
 #![no_std]
 #![no_main]